Remove todo-comments
The plugin uses the vim match function multiple times for each line on
the screen in the WinScrolled autocmd. The match function does not cache
the compiled regex and thus a considerable time is spent on compiling
regexps causing lag.
nvim-treesitter implemented TODO/FIXME highlighting some time ago, so
todo-comments is not necessary anymore.